Improve selection with ShapeCast2D instead of Area2D
Detecting selection is now forced instantly instead of waiting for collision or signals to happen.

Separate input to PlayerInput and the mouse selection logic to MouseDetector (previously SelectionDetector).

Only click to select now. Stopped using the dragging mode for selection. But the code for it is still there so it can easily be added back.